apps/payment-simulator/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
prometheus-client==0.20.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-logging==0.44b0
structlog==24.1.0
